pyproject.toml
[project]
name = "fastapi-vision-service"
version = "0.1.0"
description = "FastAPI Vision Service with Text-to-Image and Vision-Language capabilities"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "python-multipart>=0.0.12",
    "pillow>=11.0.0",
    "httpx>=0.27.0",
    "torch>=2.5.0",
    "torchvision>=0.20.0",
    "transformers>=4.46.0,<4.50.0",
    "diffusers>=0.31.0,<0.32.0",
    "accelerate>=1.1.0",
    "safetensors>=0.4.5",
    "sentencepiece>=0.2.0",
    "protobuf>=5.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
    "pre-commit>=4.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
